pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "market-making-rl-bot"
version = "0.1.0"
description = "Simulation-first reinforcement-learning market-making research scaffold using Binance market data semantics."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Tim" }]
dependencies = [
    "numpy>=1.24",
    "pydantic>=2.6",
    "httpx>=0.26",
    "websockets>=12.0",
    "orjson>=3.9",
    "gymnasium>=0.29",
    "torch>=2.1",
    "rich>=13.7",
    "typer>=0.9",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.8",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
